@@ -1,38 +1,39 @@
 {
-  "name": "test-tracker",
-  "version": "1.0.0",
+  "name": "v3_test_tracker",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "rollup": "rollup",
-    "build": "rollup -c",
-    "dev": "rollup -c -w ",
-    "env:dev": "env-cmd npm run dev",
-    "start": "sirv public --single -p 8080",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
+    "build": "npx eslint ./src && vite build",
+    "dev": "vite dev --port=8080",
+    "preview": "vite preview",
+    "serve": "vite sirv public --port=8080",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "lint": "yarn format && npx eslint --fix --ext .js,.ts,.svelte ./src",
+    "format": "prettier --ignore-path .gitignore --write \"**/*.+(js|ts|json|svelte)\""
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-node-resolve": "^11.0.0",
-    "@rollup/plugin-typescript": "^8.0.0",
-    "@tsconfig/svelte": "^2.0.0",
-    "@types/uuid": "^9.0.0",
-    "rollup": "^2.3.4",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "snarkdown": "^2.0.0",
-    "svelte": "^3.0.0",
-    "svelte-check": "^2.0.0",
-    "svelte-navigator": "^3.1.5",
-    "svelte-preprocess": "^4.0.0",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.0"
+    "@sveltejs/vite-plugin-svelte": "^2.0.3",
+    "@tsconfig/svelte": "^3.0.0",
+    "@types/uuid": "^9.0.1",
+    "eslint": "^8.35.0",
+    "prettier-plugin-svelte": "^2.9.0",
+    "svelte": "^3.55.1",
+    "svelte-check": "^2.10.3",
+    "tslib": "^2.5.0",
+    "typescript": "^4.9.3",
+    "vite": "^4.1.0"
   },
   "dependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.54.1",
+    "@typescript-eslint/parser": "^5.54.1",
     "axios": "^0.27.2",
-    "sirv-cli": "^2.0.0",
+    "eslint-config-prettier": "^8.7.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "prettier": "^2.8.4",
     "svelte-github-login": "^1.3.1",
+    "svelte-navigator": "^3.2.2",
     "sveltestrap": "^5.10.0",
     "uuid": "^9.0.0"
   }
